tạm ổn, cần tối ưu lại đăng nhập, đăng ký và tìm kiểm tra từ home

+++ rest-api/DictionaryAppForIT/DTO/WordHistoryService.cs
@@ -22,21 +22,28 @@ namespace DictionaryAppForIT.DTO
         {
             try
             {
-                HttpResponseMessage response = await client.GetAsync(apiUrl + $"get-word-lookup-history/{Class_TaiKhoan.IdTaiKhoan}");
-
-                string responseContent = await response.Content.ReadAsStringAsync();
-
-                var apiResponse = JsonConvert.DeserializeObject<ApiResponse<List<WordLookupHistory>>>(responseContent);
-
-
-                if (apiResponse.Status && apiResponse.Data != null)
+                if (Class_TaiKhoan.authentication(client))
                 {
-                    var historyData = apiResponse.Data;
-                    return historyData;
+                    HttpResponseMessage response = await client.GetAsync(apiUrl + $"get-word-lookup-history/{Class_TaiKhoan.IdTaiKhoan}");
+
+                    string responseContent = await response.Content.ReadAsStringAsync();
+
+                    var apiResponse = JsonConvert.DeserializeObject<ApiResponse<List<WordLookupHistory>>>(responseContent);
+
+
+                    if (apiResponse.Status && apiResponse.Data != null)
+                    {
+                        var historyData = apiResponse.Data;
+                        return historyData;
+                    }
+                    else
+                    {
+                        RJMessageBox.Show(apiResponse.Message);
+                        return null;
+                    }
                 }
                 else
                 {
-                    RJMessageBox.Show(apiResponse.Message);
                     return null;
                 }
             }
@@ -50,20 +57,22 @@ namespace DictionaryAppForIT.DTO
         {
             try
             {
-                HttpResponseMessage response = await client.DeleteAsync(apiUrl + $"delete-by-id-word-lookup-history/{Class_TaiKhoan.IdTaiKhoan}/{item.Index}");
-
-                string responseContent = await response.Content.ReadAsStringAsync();
-
-                var apiResponse = JsonConvert.DeserializeObject<ApiResponse<object>>(responseContent);
-
-                if (apiResponse.Status && apiResponse.Data != null)
+                if (Class_TaiKhoan.authentication(client))
                 {
-                    RJMessageBox.Show(apiResponse.Message);
-                }
-                else
-                {
-                    string message = apiResponse.Message;
-                    RJMessageBox.Show(message);
+                    HttpResponseMessage response = await client.DeleteAsync(apiUrl + $"delete-by-id-word-lookup-history/{Class_TaiKhoan.IdTaiKhoan}/{item.Index}");
+
+                    string responseContent = await response.Content.ReadAsStringAsync();
+
+                    var apiResponse = JsonConvert.DeserializeObject<ApiResponse<object>>(responseContent);
+
+                    //if (apiResponse.Status && apiResponse.Data != null)
+                    //{
+                    //    RJMessageBox.Show(apiResponse.Message);
+                    //}
+                    //else
+                    //{
+                    //    RJMessageBox.Show(apiResponse.Message);
+                    //}
                 }
             }
             catch (Exception ex)
@@ -75,20 +84,22 @@ namespace DictionaryAppForIT.DTO
         {
             try
             {
-                HttpResponseMessage response = await client.DeleteAsync(apiUrl + $"delete-translate-by-id/{Class_TaiKhoan.IdTaiKhoan}/{item.Index}");
-
-                string responseContent = await response.Content.ReadAsStringAsync();
-
-                var apiResponse = JsonConvert.DeserializeObject<ApiResponse<object>>(responseContent);
-
-                if (apiResponse.Status && apiResponse.Data != null)
+                if (Class_TaiKhoan.authentication(client))
                 {
-                    RJMessageBox.Show(apiResponse.Message);
-                }
-                else
-                {
-                    string message = apiResponse.Message;
-                    RJMessageBox.Show(message);
+                    HttpResponseMessage response = await client.DeleteAsync(apiUrl + $"delete-translate-by-id/{Class_TaiKhoan.IdTaiKhoan}/{item.Index}");
+
+                    string responseContent = await response.Content.ReadAsStringAsync();
+
+                    var apiResponse = JsonConvert.DeserializeObject<ApiResponse<object>>(responseContent);
+
+                    //if (apiResponse.Status && apiResponse.Data != null)
+                    //{
+                    //    RJMessageBox.Show(apiResponse.Message);
+                    //}
+                    //else
+                    //{
+                    //    RJMessageBox.Show(apiResponse.Message);
+                    //}
                 }
             }
             catch (Exception ex)
